requirements.txt
# Dependências do KaBot
discord.py>=2.3.0
python-dotenv
requests
google-genai # Substituindo google-generativeai para a biblioteca oficial mais recente
google-api-python-client
supabase
